Get remote .conf files
With this patch it is possible to get remote .conf files using http get. The url provided must return a json array containing the name of the files as strings and the content of the files as byte arrays.
Example: 

[
{
"name":"telconfig.conf",
"content:"byte[]"
},
{
"name":"telconfig1.conf",
"content":"byte1[]"
}
]

The files will be copied to the path specified in flag config-directory.