Get remote plugins
This patch allows to get remote .so plugins using http get. The url provided must return a json array containing the name of the files as strings and the content of the files as byte arrays.
Example: 

[
{
"name":"plugin.so",
"content:"byte[]"
},
{
"name":"plugin1.so",
"content":"byte1[]"
}
]

The files will be copied to the path specified in flag external-plugins.